fix(mac-crafter): Extract team identifier verifier to separate file
